+++ src/main/java/backend/wal/wal/todaywal/application/port/in/TodayWalSettingUseCase.java
@@ -1,6 +1,5 @@
-package backend.wal.wal.todaywal.application.port;
+package backend.wal.wal.todaywal.application.port.in;
 
-import backend.wal.wal.common.domain.WalCategoryType;
 import backend.wal.wal.common.domain.WalTimeType;
 import backend.wal.wal.nextwal.domain.NextWals;
 
@@ -9,10 +8,4 @@ import java.util.Set;
 public interface TodayWalSettingUseCase {
 
     void setTodayWals(Set<WalTimeType> timeTypes, Long userId, NextWals nextWals);
-
-    void registerTodayWal(Long userId, WalTimeType timeType, WalCategoryType categoryType, String content);
-
-    void deleteTodayWalsByWillCancelAfterNow(Set<WalTimeType> willCancelAfterNow, Long userId);
-
-    Set<WalTimeType> extractAfterNow(Set<WalCategoryType> canceledCategoryTypes, Long userId);
 }
