+++ src/main/java/backend/wal/wal/todaywal/application/port/in/UpdateTodayWalUseCase.java
@@ -1,6 +1,7 @@
-package backend.wal.wal.todaywal.application.port;
+package backend.wal.wal.todaywal.application.port.in;
 
 import backend.wal.wal.common.domain.WalCategoryType;
+import backend.wal.wal.common.domain.WalTimeType;
 
 import java.util.Set;
 
@@ -8,5 +9,8 @@ public interface UpdateTodayWalUseCase {
 
     void updateShowStatus(Long todayWalId, Long userId);
 
-    void updateTodayWalByCanceledCategoryTypesAndUserId(Set<WalCategoryType> canceledCategoryTypes, Long userId);
+    Set<WalTimeType> updateTodayWalByCanceledCategoryTypesAndUserId(Set<WalCategoryType> canceledCategoryTypes,
+                                                                    Long userId);
+
+    void deleteTodayWalsByWillCancelAfterNow(Set<WalTimeType> willCancelAfterNow, Long userId);
 }
